before delete add  check group exist

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/AlertGroupService.java
@@ -194,6 +194,12 @@ public class AlertGroupService extends BaseService{
             return result;
         }
 
+        AlertGroup alertGroup = alertGroupMapper.selectById(id);
+        if (alertGroup == null) {
+            putMsg(result, Status.ALERT_GROUP_NOT_EXIST);
+            return result;
+        }
+
         userAlertGroupMapper.deleteByAlertgroupId(id);
         alertGroupMapper.deleteById(id);
         putMsg(result, Status.SUCCESS);
